Add files via upload
One of the most widely used media and video streaming services is Amazon Prime. They have about 200 million subscribers worldwide as of mid-2021, and they have about 10,000 movies and TV series available on their site. Listings of every film and television program that is accessible on Amazon Prime are included in this tabular dataset, together with information about the actors, directors, ratings, year of release, running time, and other statistics.